requirements.txt
pillow
requests
pandas
opencv-python
numpy
torch
tkinter
random
mathplotlib